Registered more devices into the io group
This way of interpreting the opcode while executing makes adding new
functions far simpler and easy to work with. I got the idea from how
MEMC works on the acorn archimedes being write only. I am allowing reads
to be a way to get direct values as well. The operations can be blocking
or non blocking depending on what is being requested. I also have
16-million available operation combinations. All arguments are passed
to the transaction "cache line". Thus we can easily use it as a way to
perform some sort of extra action.

This idea also applies to more than just the current design as well. It
does require a rewrite of the i960 firmware but at this point I don't
care. It just makes the entire process cleaner