Move write operations of the cache to a separate function
Right now, we just want to get the write request off the bus as fast as
possible. We will then commit the transaction later on. If the cache
line is 16-bytes in size and we do an 8-word transfer then there is no
need to load the cache from memory. Instead, just sync the cache line
and then overwrite it with our contents. For now, we do the cache lookup
_after_ we get the data off of the bus to do it as fast as possible