Introduce what I call double direct mapped 4
It is a direct mapped implementation which holds onto four separate lines per set. We introduce two more bits between offset and tag to select one of the lines inside this set. There is still a one to one mapping but locality means that we can still hold more related data per entry.

This may do nothing but I want to see