use the simplest sd card implementation to load data off of it
Delete xmem and the sdfat libraries

I really don't need them anymore except to bootstrap the system. I was going to include a monitor but that turns out to be too much of a pain in the ass.

I removed xmem because I was using it for malloc operations which I don't need anymore