Don't use bit fields for bank accessing
Instead do the math directly, it allows me to pull the offset by setting the upper most bit of the lower half to 1 in all cases and return it (a single instruction instead of multiple)

When computing the bank id, we take the upper half, shift it by one and then or the bit we need from lower half into it. It is far faster than it was before just in sheer number of instructions. Doing a shift by 17 causes a loop to be performed :/.