Do alignment tracking on the first 16-bit word of a read
This allows me to only assign the 32-bit word once and pulse quickly
because the data is already written once. Right now, it is only if you
start on the lower 16-bits. If you start on the upper 16-bits then the
alignment does not take place. However, I am planning to implement this
as well.