Added a check to terminate early when the upper byte enable of a 16-bit value is one.
The following 16-bit combinations are what's possible:

0b00 - 16-bit op
0b10 - lower 8-bit op
0b01 - upper 8-bit op
0b11 - invalid

In all cases, when we hit lower 8-bit op, it immediately means terminate
immediately with no need to evaluate blast as it will _always_ be low.

I have implemented this idea for the later parts of a burst transaction.
Specifically for write operations. Read operations will operate the same
as always.

I also realized that the later parts of a burst transaction will always
have the lower 8-bits of the 16-bit word be applied. There is no need to
evaluate that. It only applies to the first word in the transaction to
figure out where the first "zero" is located. Then we keep going until
we find the first "one". It is impossible with the i960 instruction set
to have discontiguous spans of bytes. For example, 0b1010 is impossible
to get as a series of byte enable bits. However, 0b1001 is. So this idea
also applies to 32-bit bus variants just with more states to care about.