Reduce operating time by inverting the order of byte enable bit checks on transaction startup.
Check the upper bit instead of the lower bit first. By doing that, you
actually eliminate a blast check all together. If the higher bit is high
then it means we are actually done and don't need to check the lower
byte enable bit. It will just be captured from the bus and then we break
out.

When higher BE is low, then we perform the assignment and then check BE0
to see if we should update it as well. Generally, this will also be true
but it is important to check it since this is a write operation.