The XBUS is now reserved for future "snooping" capabilities
I am slowly migrating to a design where the i960 directly accesses its
memory and the 2560 acts as a signaller and also citizen on the bus. The
XBUS allows it to act as a bus master in this case. The 32k window is
effective for this as well. The CH351 I allocated for this is very
useful for this exact purpose.

The only part that will need to be redesigned is the capture dimm
because it requires me to install buffers to insulate the 2560 from the
i960's bus when it is not in use. Currently, it is responsible for
serving data to the i960 from memory itself. A relatively slow process
that costs around 50-60 i960 cycles for a single 64-bit transfer
regardless of direction.