Get rid of the logic for directly connecting data lines to the 2560
After doing some offline testing, I found that directly connecting the
data lines to the microcontroller did not yield as much of a benefit as
I thought. I found that the loss of pins plus overall upheaval of the
design was not worth it. Especially since my goal is to migrate towards
the i960 being able to talk to external devices directly.