Make sure that we make the ibus the default for both read and write operations
I realized that I can expand the size of the IBUS to 64k * 16k from 256 * 16k.
This increases the amount of addressible memory from 4 megabytes to 1 gigabyte.
This is a very large amount of SRAM that I probably would never use
directly. However, it does give me the ability to control how many pins
I allocate towards bank selection as well.

The biggest part about all of this is that I can use this to start
migrating towards the overall goal of having the i960 get its data
directly!