Only read the parts of the address that we actually care about
Since we now do decoding of the upper address byte via external
hardware, there is no reason to read in the complete address.

When dealing with io operations, it is the lower 24-bits that we
actually need to read in.

When it is an IBUS request (which mirrors the same 4 megabyte window
through out the entire 32-bit memory space not consumed by the IO space)
we only need to read the lower half of the address. This shaves off
quite a bit of time. We are now down to ~54 cycles from 58!