There is now a strange hang that I have traced back to the AVR code
It is expressed in both gcc 7 and gcc 11. The serial console does not
fully print out messages if the buffer is too large. If I reduce the
size of the buffer then I get more and more text out of it but still a
"hang". I am not sure what the actual issue is here but it is quite
strange overall.