Register a new bus decoder gal
It is designed to carve the 32-bit memory space up into several simple
parts. It provides for 16 megabytes of flash at the start of ram and the
end of ram as well (starting at 0xFE000000 to be more exact). In the case of the Sx and Kx chips,
we mirror the flash to create a common boot environment. We connect a
single microcontroller pin to the MIRROR pin to enable or disable this
functionality. This pin can also be connected to the CFG pins as well to
provide automatic configuration too! Not sure yet!