Use auto generation for the Timer interface since it is a special case
In most other situations, accessing memory like expected would be just
fine. However, the macro expansion causes problems and confuses the
compiler. Thus we need to do template specialization. I still want it to
be a discrete instance but this is the best that can be done for the
TimerInterface as is